requirements.txt
ultralytics
opencv-python
numpy
pillow
python-dotenv
